Replaced the debugStore property with a store property that takes the internal name of the class to use as the Store.  Default is com/surelogic/_flashlight/Store.  But this allows substitution of DebugStore, and a new EmptyStore.  Affects Configuration.java as well as the ant task FLInstrument.
Rewrote Entities.escape(String) to be much more efficient.  Previous version did a lot of inserting which causes repeated copying inside the StringBuilder.  New version just appends substrings to a StringBuilder that has already been sized to (hopefully) avoid any extra copying at all.

Added an abstract getWithin() method to KeyField.  THis returns the PhantomRefernece that the object compares against in the isWithin() method.  This is necessary for the improvements made to Refinery.  Affects the concrete subclasses KeyFieldInstance and KeyFieldStatic.

Made improvements to the data structures used by Refinery to maintain O(1) operations:
* Replaced f_fieldToThread (a Map<KeyField, PhantomReference) with f_objToFieldToThread, a Map<PhantomReference, Map<KeyField, PhantomReference>>.  While this requires to operations to put or get a Field->Thread mapping, it enables O(1) removal of all the fields from a particular object.  Previously this required iterating over the entire keyset.
* Replaced f_sharedField (a List<KeyField>) with a f_objToSharedFields, a Map<PhantomReference, Set<KeyField>>.  Again, trade of making normal lookups slightly more expensive, with making removing all the fields from a particular object a constant operation instead of a linear one.

Changed FlashlightRuntimeSupport to cache the results of getField().  Added a caching getClass() method to replace the use of Class.forName() in the instrumentaiton.  Had to update the instrumentation to use the new method instead of Class.forName().



git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/flashlight-instrumentation-5.0@9204 d02ae147-ee12-4a35-9452-91c89e8b2592