Added some label remapping.  In particular, now that we include the after lock acquisition method call at the start of the try-block that follows a monitorenter, we had a new problem.  The label used to mark the start of the try-block can be used a jump target.  When this happened, the stack was not set up appropriately for the after lock acquisition call, and the bytecode verifier complained.  Also, we don't want to run that method again.  So I need to add a new label that follows the instrumentation to be used for the jump target.
What we actually do is create a new label for the start of the try-block and and output the original label after the instrumentation.  That way I don't have to fix any jump operations.  I only have to modify forwarded calls to visitTryCatchBlock() and forwarded calls to visitLabel().  I am helped a lot by the fact taht visitTryCatchBlock() must appear before any of the labels it uses are visited.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/flashlight-instrumentation-5.0@9591 d02ae147-ee12-4a35-9452-91c89e8b2592