Fixed bug 1557.  The solution to this problem is to use the type of the "receiver" argument of the access method as the type of the "receiver" argument for the wrapper method.
We assume that the compilers are following the Sun guidelines and generating methods called "access$XXX".  In this case, we look for static synthetic methods whose name begins with "access$".  This is identified in the constructor of FlashlightMethodRewriter.  For these methods we record the type of the first parameter (if any); we call this the receiver type.  When instrumenting a virtual method call, we now take note of this receiver type.  If it exists, we use it instead of the owner type from the method instruction as the type of the first (receiver) parameter of wrapper method.  In this case we also expand the name of the wrapper method to include the receiver type so as to avoid method lookup problems with other wrapper methods that may invoke the same method but use a different receiver type.  The template is "flashlight$receiver_type$owner_type$original_method_name$virtualWrapper"

git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/flashlight-instrumentation-5.0@9997 d02ae147-ee12-4a35-9452-91c89e8b2592