Created StoreDelegate to contain the FL_OFF field and the phantom reference creation methods.  The Store, DebugStore, and EmptyStore classes now delegate to this class to share the implementation of the reference creation methods.  Changed IdObject to invoke StoreDelegate methods.  It cannot refernece the Store class directly because this causes serious problems if the rest of the instrumentation is using DebugStore or EmptyStore.
Updated DebugStore to create its own version of the standard output stream.  It cannot use System.out becuase this field can be reset to point to a different stream.  If that stream is implemented using a class that happens to be instrumented, then DebugStore will enter a recursive loop and blow the stack.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/flashlight-instrumentation-5.0@10646 d02ae147-ee12-4a35-9452-91c89e8b2592