Updated the Instrument ant task to be compatible with the latest changes to the RewriteManager that require the complete classpath of the instrumented application to be known.  This is so the class information can be availabe for correcting java 6 stack map frames.
Did this by adding bootclasspath, extdirs, and libraries attributes/elements to the task.  These are processes as ANT Path elements, and have all the features of ANT path-like entities.  A sanity check makes sure that bootclasspath only refers to files, and extdirs only refers to directories.

Really we just need the libraries attribute, but separating out the bootclasspath and extdirs attributes makes sense because it is how the JAVA JVM does things.  This way we can derive the bootclasspath and extdirs from the run-time system if they are not specified.



git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/flashlight-ant@11958 d02ae147-ee12-4a35-9452-91c89e8b2592