ASSIGNED - bug 1575: Update Flashlight instrumentation to selectively not instrument fields https://fluid.surelogic.com/cgi-bin/bugzilla3/show_bug.cgi?id=1575
Instrumentation and Ant task updated to support this.   The Eclipse UI still needs to be updated.

Can be set using the properties com.surelogic._flashlight.rewriter.filter.fields and com.surelogic._flashlight.rewriter.filter.fields.inPackages.  The first must be one of "none", "declaration", or "use".  The second is a list of packages (internal names with '/', not with '.').

git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/flashlight-instrumentation-5.0@12300 d02ae147-ee12-4a35-9452-91c89e8b2592