Removed the fieldAccess() method from Store.  It's been deprecated for some time now.  It was only around to keep various JUnit Test programs happy.  I have simply commented out the calls in the JUnit tests.  Those tests are now broken, but they were irrelevant at this point any way.  This also allowed me to get rid of the SrcLoc class.
Fixed the previous "missing field declaration" error that occurs when accessing fields delcared in classes that aren't instrumented, again.  Had to update the {put, get}{field, static} rewriting to push the Class object of the class that DECLARES the field being accessed, and not the Class object of the owner class provided by the bytecode.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/flashlight-instrumentation-5.0@12317 d02ae147-ee12-4a35-9452-91c89e8b2592