Updated so that a new Store.classInit() method is called when a class initializer begins and ends execution.  The STore.classInit() implementation needs to be filled in, write now it just contains a println() call.
Nathan needs this to do lock set analysis on static fields.  He needs to be able to determine if the initializer for class C is currently in the stack trace when a static field declared in C is accessed.  Such fields are to be ignored for locking purposes because we can assume serialized access at this point in the code.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/flashlight-instrumentation-5.0@12501 d02ae147-ee12-4a35-9452-91c89e8b2592