Updated Flaslight to use ASM-3.2.
Added some debugging cability to ExceptionHandlerReorderingMethodAadapter.

Fixed FlashlightMethodRewriter.visitLabel() to avoid triggering a bug in ASM.  ASM throws an exception if you have a label preceeding a JSR opcode that does not have a forward reference.  So I have to output a bogus GOTO statement to force a forward reference to exist.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/flashlight-developer-tools@12885 d02ae147-ee12-4a35-9452-91c89e8b2592