Had a problem instrumented Vuze.  Turned out method wrappers being called by access$ methods were all messed up.  This was caused by a misapplication of the fixed intended to fix bug 1557.  But it seems that bug 1557 was already separately readdressed in March 2009 unintentially:
It seems that back on 13 March 2009, I encountered a similar problem with
invokevirtual that appeared outside of an access$ method.  See the notes on
Version 11914 of FlashlightMethodRewriter.java in SVN:

"Updated the method rewriter to handle a funny case of INVOKEVIRTUAL that we
discovered in a jar file downloaded from the net.  Basically, class C extends
S.  Code in C.m() that invokes method n() on the reciever where n() is declared
in S and not overridden in C is generated as INVOKEVIRTUAL S.n(), whereas
normally it would be generated as INVOKEVIRTUAL C.n().  This makes my static
wrapper methods for the emthod call illegal when n() is not visible in the
static context (such as when n() is protected and S and C are in different
packages).  

"The fix is to use in-place instrumentation when the method being instrumented
is in class C, but the owner of the method being called is S, the superclass of
C.  There are two prolbems with this:
"(1) The same trick could be played when n() is defined in S', a superclass of
S.  I don't have access to the full class hierarchy, so I cannot test for this.
"(2) We don't distinguish the case when the class is not this.n(), but x.n(),
and the static type of x really is S.

This fix actually causes the original fix to not even be used.  But the we have
run into a case where the original fix is being triggered when it is not
appropriate and messing things up.  I think the best thing to do is to remove
the original attempt to fix that that is limited to access methods.

git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/flashlight-client-eclipse@12893 d02ae147-ee12-4a35-9452-91c89e8b2592