RESOLVED - bug 1643: Update flashlight to track assignments made to fields during construction/initialization https://fluid.surelogic.com/cgi-bin/bugzilla3/show_bug.cgi?id=1643
Instrumentation now issues calls to new Store.instanceFieldInit() and Store.staticFieldInit() methods when it encounters putfield and putstatic operations, respectively, in <init> and <clinit> methods, respectively.  instanceFieldInit() is only called if the field is object-valued and the receiver is the object being constructed. staticFieldInit() is only called if the field is object-valued and the field is decalred in the class that contains the <clinit> method.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/flashlight-instrumentation-5.0@14080 d02ae147-ee12-4a35-9452-91c89e8b2592