Fixed up RewriteManager to load the happens-before config file.
Things seem to work the way I intended them to, but I realized what I'm
doing is probably not correct.  This only works if the static types of
the receivers match the config file information exactly.  It won't work,
for example, if the static type is a super type of an interesting class
even if the actual dynamic type of the referenced object is an
interesting class.

If the static type of the receiver is X, and we are invoking method
m(x,y), and there is a happens-before method m(x,y) in type Y such that
X is an ancestor of Y, then the method call MIGHT be interesting.  We
have to check at runtime the actual type.

If the static type of the receiver is Z, where Y is an ancestor of
Z, we MUST instrument because they are definitely interesting.

So we have three scenarios
1 - Y.m is called directly.  In this case everything is fine
2 - Z extends Y, and Z.m is called.  We can find these, everything is
still fine
3 - Y extends X, and X.m is called.  We can detect these, but only at
runtime would we know if X is really a subclass of Y

The rewriter can statically detect the 3 cases.  Cases 1 and 2 are
really the same.  I think I should add 2 params to each of
happensBeforeObject(), happensBeforeThread(), and
happensBeforeCollection():
  (1) a flag that if true indicates cases 1 and 2, and if false
      indicates case 3/
  (2) a string that is the class name Y that is used when the flag is
      false.  This can be null when the flag is true.
      
The Store method implementation can, when the flag is false, use the
class name to get the Class object for Y, and then get the class object
for the method receiver (which is always the first argument of each
happensBefore method) and test whether the type of the receiver is
assignable to Y.  (Method Class.isAssignableFrom())



Former-commit-id: 557dd3dc8314762dcb4479c29ed3a2bebbd9abeb